runqemu-export-rootfs: refactor to allow more protocols than NFS
Despite the generic name, the script is pretty much tailored in its
parameters, variable names and functionality to NFS.

In preparation for adding 9P export support via diod, refactor it, so
the generic option parsing and command handling are separated from the
NFS specific bits.

This patches introduces functional changes for existing users:
NFS_INSTANCE has been renamed to INSTANCE, and MOUNTD_PORT and NFSD_PORT
have been combined into a single PORT. It's ok to set them to the same
port and the only in-tree user, scripts/runqemu, does this already, so
it shouldn't cause much issue to just remove the ability to set them
differently when calling the script.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>